Move migration popup to deploy flows: intercept rocket icon and Deploy button for staging
